requirements.txt
# Agent 项目依赖
openai>=1.0.0
python-dotenv>=1.0.0
